+++ security-groups.tf
@@ -1,11 +1,11 @@
 resource "aws_security_group" "alb" {
-  name   = "${var.project_name}-load-balancer-security-group"
+  name   = "${var.project_name}-${var.environment}-load-balancer-security-group"
   vpc_id = var.vpc_id
 
   ingress {
     protocol    = "tcp"
-    from_port   = var.container_port
-    to_port     = var.container_port
+    from_port   = var.alb_ingress_port
+    to_port     = var.alb_ingress_port
     cidr_blocks = ["0.0.0.0/0"]
   }
 
@@ -18,13 +18,13 @@ resource "aws_security_group" "alb" {
 }
 
 resource "aws_security_group" "ecs_tasks" {
-  name   = "${var.project_name}-ecs-tasks-security-group"
+  name   = "${var.project_name}-${var.environment}-ecs-tasks-security-group"
   vpc_id = var.vpc_id
 
   ingress {
     protocol        = "tcp"
-    from_port       = var.container_port
-    to_port         = var.container_port
+    from_port       = var.host_port
+    to_port         = var.host_port
     security_groups = [aws_security_group.alb.id]
   }