+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+- Add capacity provider support to support spot instances for cost savings
+
 ## [0.0.2] - 2023-04-24
 
 ### Fixed
@@ -11,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Adjust ALB variables
 - Adjust ECS variables
 - Add IAM actions variable
+- Forced HTTPS on the ALB and added ACM certificate
 
 ## [0.0.1] - 2023-04-18